+++ sbin/searchCorrespondingImages.sh
@@ -38,12 +38,12 @@ if [ "${USESIMILARITYINDEX}" != "" ]; then
     cd ${CWD}
     echo "now: check images and image-index-result with database: ${SEARCHDIR}"
     cd ${MYTB}
-    node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.json --command mediaManager --action findCorrespondingTourDocRecordsForMedia --importDir $SEARCHDIR  --additionalMappingsFile $SEARCHDIR/findFilesInLireIndex.json --debug true > $SEARCHDIR/findFilesInDb.tmp && sed -e '/DONE - command finished/,$d' $SEARCHDIR/findFilesInDb.tmp > $SEARCHDIR/findFilesInDb.json
+    node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.dev.json --command mediaManager --action findCorrespondingTourDocRecordsForMedia --importDir $SEARCHDIR  --additionalMappingsFile $SEARCHDIR/findFilesInLireIndex.json --debug true > $SEARCHDIR/findFilesInDb.tmp && sed -e '/DONE - command finished/,$d' $SEARCHDIR/findFilesInDb.tmp > $SEARCHDIR/findFilesInDb.json
     cd ${CWD}
 else
     echo "now: check images with database: ${SEARCHDIR}"
     cd ${MYTB}
-    node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.json --command mediaManager --action findCorrespondingTourDocRecordsForMedia --importDir $SEARCHDIR  --debug true > $SEARCHDIR/findFilesInDb.tmp && sed -e '/DONE - command finished/,$d' $SEARCHDIR/findFilesInDb.tmp > $SEARCHDIR/findFilesInDb.json
+    node dist/backend/serverAdmin.js -c ${CONFIG_BASEDIR}backend.dev.json --command mediaManager --action findCorrespondingTourDocRecordsForMedia --importDir $SEARCHDIR  --debug true > $SEARCHDIR/findFilesInDb.tmp && sed -e '/DONE - command finished/,$d' $SEARCHDIR/findFilesInDb.tmp > $SEARCHDIR/findFilesInDb.json
     cd ${CWD}
 fi