+++ src/backend/commands/tdoc-loader.command.ts
@@ -14,7 +14,12 @@ export class TourDocLoaderCommand implements AbstractCommand {
     public process(argv): Promise<any> {
         const typeOrder = ['location', 'news', 'trip', 'route', 'track', 'image', 'video'];
 
-        const filePathConfigJson = argv['c'] || argv['backend'] || 'config/backend.json';
+        const filePathConfigJson = argv['c'] || argv['backend'];
+        if (filePathConfigJson === undefined) {
+            console.error('ERROR - parameters required backendConfig: "-c | --backend"');
+            process.exit(-1);
+        }
+
         const serverConfig = {
             backendConfig: JSON.parse(fs.readFileSync(filePathConfigJson, { encoding: 'utf8' })),
             readOnly: false