+++ src/frontendserver/frontendserverAngular.ts
@@ -23,9 +23,13 @@ const debug = argv['debug'] || false;
 const staticFolder = join(process.cwd(), 'dist/static');
 const distProfile = 'DIST_PROFILE';
 const distServerProfile = 'DIST_SERVER_PROFILE';
-const filePathConfigJson = argv['frontend'] || 'config/frontend.json';
-const filePathBackendConfigJson = argv['c'] || argv['backend'] || 'config/backend.json';
-const filePathFirewallConfigJson = argv['f'] || argv['firewall'] || 'config/firewall.json';
+const filePathConfigJson = argv['frontend'];
+const filePathBackendConfigJson = argv['c'] || argv['backend'];
+const filePathFirewallConfigJson = argv['f'] || argv['firewall'];
+if (filePathConfigJson === undefined || filePathBackendConfigJson === undefined || filePathFirewallConfigJson === undefined) {
+    console.error('ERROR - parameters required frontendConfig:  "--frontend" backendConfig: "-c | --backend" firewallConfig: "-f | --firewall"');
+    process.exit(-1);
+}
 
 export interface ServerConfig {
     filePathErrorDocs: string;